[CHORE] MemoryViewController를 Presentation으로 이동(#563)
